third_party/blink/web_tests/fast/parser/scriptexec-during-parserRemoveChild-expected.txt

Ensure that iframe onunload event handler triggered during reparenting that modifies the reparented DOM doesn't corrupt DOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iframe is iframe.parentNode.firstChild
PASS parentNode is parentNode.parentNode.firstChild
PASS successfullyParsed is true

TEST COMPLETE
